add batch mode for HD
