+++ src/EditorHeader.tsx
@@ -1,17 +1,17 @@
-import { ArrowLeftIcon } from '@heroicons/react/outline'
+import { ArrowLeftIcon, DownloadIcon, EyeIcon } from '@heroicons/react/outline'
 import { useWindowSize } from 'react-use'
-import { useUser } from './adapters/user'
 import Button from './components/Button'
-import Logo from './components/Logo'
-import LogoPro from './components/LogoPro'
 import Menu from './components/Menu'
 import Toggle from './components/Toggle'
+import { useEditor } from './context/EditorContext'
 
 interface EditorHeaderProps {
   onBack: () => void
   useHD: boolean
   setUseHD: (useHD: boolean) => void
   setShowUpgrade: (showUpgrade: boolean) => void
+  setShowOriginal: (showOriginal: boolean) => void
+  setShowSeparator: (showSeparator: boolean) => void
 }
 
 export default function EditorHeader({
@@ -19,26 +19,47 @@ export default function EditorHeader({
   useHD,
   setUseHD,
   setShowUpgrade,
+  setShowOriginal,
+  setShowSeparator,
 }: EditorHeaderProps) {
-  const user = useUser()
   const windowSize = useWindowSize()
-
-  function getLogo() {
-    if (user?.isPro()) {
-      return <LogoPro className={['h-12'].join(' ')} />
-    }
-    return <Logo className={['h-12'].join(' ')} />
-  }
+  const editor = useEditor()
 
   return (
     <header className="relative z-10 flex sm:px-5 pt-3 justify-between items-center sm:items-start">
       <Button icon={<ArrowLeftIcon className="w-6 h-6" />} onClick={onBack}>
         {windowSize.width > 640 ? 'Start new' : undefined}
       </Button>
-      {windowSize.width > 640 ? getLogo() : <></>}
 
       <div className="flex space-x-4">
-        <Toggle label="HD" enabled={useHD} setEnabled={setUseHD} />
+        <div className="mr-4 flex items-center">
+          <Toggle label="HD" enabled={useHD} setEnabled={setUseHD} />
+        </div>
+        {editor.renders.length ? (
+          <>
+            <Button
+              icon={<EyeIcon className="w-6 h-6" />}
+              onDown={ev => {
+                ev.preventDefault()
+                setShowOriginal(true)
+                setShowSeparator(true)
+              }}
+              onUp={() => {
+                setShowOriginal(false)
+                setTimeout(() => setShowSeparator(false), 300)
+              }}
+            />
+            <Button
+              primary
+              icon={<DownloadIcon className="w-6 h-6" />}
+              onClick={editor.download}
+            >
+              {windowSize.width > 640 ? 'Download' : undefined}
+            </Button>
+          </>
+        ) : (
+          <></>
+        )}
         <Menu onUpgrade={() => setShowUpgrade(true)} />
       </div>
     </header>
